Move blocking from delegate result to method
`SplitScannerCoordinatorDelegate.didScanBarcode` once again returns just
a `ScanResult`, instead of `(result: ScanResult, blocking: Bool)`. To
block the scanner, call
`splitScannerCoordinator.blockScanner(withMessage: barcode)`.

This allows the elimination of a race condition when unblocking the
scanner at the completion of an asynchronous request. If the
asynchronous request finished and attempted to unblock the scanner
before the scanner blocked itself, the scanner would remain blocked
indefinitely.