Add new `pending` case to `ScanResult`
This commit adds a new `pending` option to `ScanResult`. This is
useful for cases where a network request needs to be performed
before it can be determined whether a result should be a
`success`, `warning`, or `error`. A `pending` result is not added
to the table. A new function is also introduced to allow
manually adding a result to the table, so that when the
network task is complete, the appropriate result can be
displayed.