Add ability to manually pop the root coordinator
There are instances in which it is convenient to be able to manually dismiss
the SplitScannerCoordinator. However, currently, dismissing the coordinator
from its presenting controller causes a retain cycle, since the
SplitScannerCoordinator is never popped from its root coordinator. This
allows for popping the coordinator from outside of itself, which subsequently
allows for safe dismissal from the presenting controller.